requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
addict=2.4.0=pypi_0
albucore=0.0.13=pypi_0
albumentations=1.3.1=pypi_0
aliyun-python-sdk-core=2.15.1=pypi_0
aliyun-python-sdk-kms=2.16.3=pypi_0
annotated-types=0.7.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
asttokens=2.4.1=pypi_0
astunparse=1.6.3=pypi_0
attrs=24.2.0=pypi_0
black=24.4.2=pypi_0
blinker=1.9.0=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.3.11=h06a4308_0
certifi=2024.6.2=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
clip=1.0=pypi_0
cloudpickle=3.0.0=pypi_0
colorama=0.4.6=pypi_0
comm=0.2.2=pypi_0
configargparse=1.7=pypi_0
contourpy=1.2.1=pypi_0
crcmod=1.7=pypi_0
cryptography=42.0.8=pypi_0
cycler=0.12.1=pypi_0
dash=2.18.2=pypi_0
dash-core-components=2.0.0=pypi_0
dash-html-components=2.0.0=pypi_0
dash-table=5.0.0=pypi_
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
detectron2=0.6=pypi_0
docker-pycreds=0.4.0=pypi_0
einops=0.8.0=pypi_0
eval-type-backport=0.2.0=pypi_0
executing=2.1.0=pypi_0
fastjsonschema=2.20.0=pypi_0
filelock=3.14.0=pypi_0
flask=3.0.3=pypi_0
flatbuffers=24.3.25=pypi_0
fonttools=4.53.0=pypi_0
fsspec=2024.6.0=pypi_0
ftfy=6.3.1=pypi_0
fvcore=0.1.5.post20221221=pypi_0
gast=0.6.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.64.1=pypi_0
h5py=3.11.0=pypi_0
huggingface-hub=0.23.4=pypi_0
hydra-core=1.3.2=pypi_0
idna=3.7=pypi_0
imageio=2.34.2=pypi_0
importlib-metadata=7.2.1=pypi_0
iopath=0.1.9=pypi_0
ipython=8.29.0=pypi_0
ipywidgets=8.1.5=pypi_0
itsdangerous=2.2.0=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.4=pypi_0
jmespath=0.10.0=pypi_0
joblib=1.4.2=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
jupyter-core=5.7.2=pypi_0
jupyterlab-widgets=3.0.13=pypi_0
kagglehub=0.3.4=pypi_0
keras=3.5.0=pypi_0
kiwisolver=1.4.5=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libclang=18.1.1=pypi_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
lightning-utilities=0.11.6=pypi_0
llvmlite=0.43.0=pypi_0
markdown=3.6=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.9.0=pypi_0
matplotlib-inline=0.1.7=pypi_0
mdurl=0.1.2=pypi_0
ml-dtypes=0.4.1=pypi_0
mmcv=2.1.0=pypi_0
mmcv-lite=2.2.0=pypi_0
mmdet=3.3.0=pypi_0
mmengine=0.10.4=pypi_0
mmyolo=0.6.0=pypi_0
model-index=0.1.11=pypi_0
mpmath=1.3.0=pypi_0
multiscaledeformableattention=1.0=pypi_0
mypy-extensions=1.0.0=pypi_0
namex=0.0.8=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
networkx=3.3=pypi_0
numba=0.60.0=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.20.5=pypi_0
nvidia-nvjitlink-cu12=12.5.40=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
omegaconf=2.3.0=pypi_0
open3d=0.18.0=pypi_0
opencv-python=4.10.0.84=pypi_0
opencv-python-headless=4.10.0.84=pypi_0
opendatalab=0.0.10=pypi_0
openmim=0.3.9=pypi_0
openssl=3.0.14=h5eee18b_0
openxlab=0.1.0=pypi_0
opt-einsum=3.3.0=pypi_0
optree=0.12.1=pypi_0
ordered-set=4.1.0=pypi_0
oss2=2.17.0=pypi_0
packaging=24.1=pypi_0
pandas=2.2.2=pypi_0
parso=0.8.4=pypi_0
pathspec=0.12.1=pypi_0
pexpect=4.9.0=pypi_0
pillow=10.3.0=pypi_0
pip=24.0=py311h06a4308_0
platformdirs=4.2.2=pypi_0
plotly=5.24.1=pypi_0
pointpillars=0.0.0=dev_0
portalocker=2.10.0=pypi_0
prettytable=3.10.0=pypi_0
prompt-toolkit=3.0.48=pypi_0
protobuf=4.25.3=pypi_0
psutil=6.1.0=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
pycocotools=2.0.8=pypi_0
pycparser=2.22=pypi_0
pycryptodome=3.20.0=pypi_0
pydantic=2.8.2=pypi_0
pydantic-core=2.20.1=pypi_0
pygments=2.18.0=pypi_0
pyparsing=3.1.2=pypi_0
pypcd4=1.1.0=pypi_0
pyquaternion=0.9.9=pypi_0
python=3.11.9=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
python-lzf=0.2.6=pypi_0
pytorch-ood=0.1.9=pypi_0
pytz=2023.4=pypi_0
pyyaml=6.0.1=pypi_0
qudida=0.0.4=pypi_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
regex=2024.5.15=pypi_0
requests=2.28.2=pypi_0
retrying=1.3.4=pypi_0
rich=13.4.2=pypi_0
rpds-py=0.21.0=pypi_0
safetensors=0.4.3=pypi_0
scikit-image=0.24.0=pypi_0
scikit-learn=1.5.1=pypi_0
scipy=1.13.1=pypi_0
seaborn=0.13.2=pypi_0
sentry-sdk=2.17.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=60.2.0=pypi_0
shapely=2.0.4=pypi_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sqlite=3.45.3=h5eee18b_0
stack-data=0.6.3=pypi_0
supervision=0.19.0=pypi_0
sympy=1.12.1=pypi_0
tabulate=0.9.0=pypi_0
tenacity=9.0.0=pypi_0
tensorboard=2.17.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorboard-logger=0.1.0=pypi_0
tensorflow-io-gcs-filesystem=0.37.1=pypi_0
termcolor=2.4.0=pypi_0
terminaltables=3.1.10=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2024.7.24=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.19.1=pypi_0
tomli=2.0.1=pypi_0
torch=2.3.1=pypi_0
torchmetrics=1.4.1=pypi_0
torchvision=0.18.1=pypi_0
tqdm=4.65.2=pypi_0
traitlets=5.14.3=pypi_0
transformers=4.41.2=pypi_0
triton=2.3.1=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.1=pypi_0
urllib3=1.26.19=pypi_0
wandb=0.18.5=pypi_0
wcwidth=0.2.13=pypi_0
werkzeug=3.0.3=pypi_0
wheel=0.43.0=py311h06a4308_0
widgetsnbextension=4.0.13=pypi_0
wrapt=1.16.0=pypi_0
xz=5.4.6=h5eee18b_1
yacs=0.1.8=pypi_0
yapf=0.40.2=pypi_0
yolo-world=0.1.0=pypi_0
zipp=3.19.2=pypi_0
zlib=1.2.13=h5eee18b_1